Conflicts:
	Library/Formula/qt.rb

+++ Library/Homebrew/formula_installer.rb
@@ -15,7 +15,7 @@ class FormulaInstaller
   def initialize ff, tab=nil
     @f = ff
     @tab = tab
-    @show_header = true
+    @show_header = false
     @ignore_deps = ARGV.ignore_deps? || ARGV.interactive?
     @install_bottle = install_bottle? ff
 
@@ -73,6 +73,12 @@ class FormulaInstaller
       needed_deps = []
       needed_reqs = []
 
+      # HACK: If readline is present in the dependency tree, it will clash
+      # with the stdlib's Readline module when the debugger is loaded
+      if f.recursive_deps.any? { |d| d.name == "readline" } and ARGV.debug?
+        ENV['HOMEBREW_NO_READLINE'] = '1'
+      end
+
       ARGV.filter_for_dependencies do
         needed_deps = f.recursive_deps.reject{ |d| d.installed? }
         needed_reqs = f.recursive_requirements.reject { |r| r.satisfied? }
@@ -99,7 +105,7 @@ class FormulaInstaller
           end
         end
         # now show header as all the deps stuff has clouded the original issue
-        show_header = true
+        @show_header = true
       end
     end
 
@@ -146,16 +152,14 @@ class FormulaInstaller
       ohai 'Caveats', f.keg_only_text
       @show_summary_heading = true
     else
-      audit_bin
-      audit_sbin
-      audit_lib
-      check_manpages
-      check_infopages
-      check_m4
+      #audit_bin
+      #audit_sbin
+      #audit_lib
+      #check_manpages
+      #check_infopages
     end
 
     keg = Keg.new(f.prefix)
-
     if keg.completion_installed? :bash
       ohai 'Caveats', <<-EOS.undent
         Bash completion has been installed to:
@@ -168,6 +172,61 @@ class FormulaInstaller
         zsh completion has been installed to:
           #{HOMEBREW_PREFIX}/share/zsh/site-functions
         EOS
+    end
+
+    if f.plist or keg.plist_installed?
+      destination = f.plist_startup ? '/Library/LaunchDaemons' \
+                                    : '~/Library/LaunchAgents'
+
+      plist_filename = f.plist_path.basename
+      plist_link = "#{destination}/#{plist_filename}"
+      plist_domain = f.plist_path.basename('.plist')
+      destination_path = Pathname.new File.expand_path destination
+      plist_path = destination_path/plist_filename
+      s = []
+
+      # we readlink because this path probably doesn't exist since caveats
+      # occurs before the link step of installation
+      if not (plist_path).file? and not (plist_path).symlink?
+        if f.plist_startup
+          s << "To have launchd start #{f.name} at startup:"
+          s << "    sudo mkdir -p #{destination}" unless destination_path.directory?
+          s << "    sudo cp -fv #{HOMEBREW_PREFIX}/opt/#{f.name}/*.plist #{destination}"
+        else
+          s << "To have launchd start #{f.name} at login:"
+          s << "    mkdir -p #{destination}" unless destination_path.directory?
+          s << "    ln -sfv #{HOMEBREW_PREFIX}/opt/#{f.name}/*.plist #{destination}"
+        end
+        s << "Then to load #{f.name} now:"
+        if f.plist_startup
+          s << "    sudo launchctl load #{plist_link}"
+        else
+          s << "    launchctl load #{plist_link}"
+        end
+        if f.plist_manual
+          s << "Or, if you don't want/need launchctl, you can just run:"
+          s << "    #{f.plist_manual}"
+        end
+      elsif Kernel.system "/bin/launchctl list #{plist_domain} &>/dev/null"
+        s << "You should reload #{f.name}:"
+        if f.plist_startup
+          s << "    sudo launchctl unload #{plist_link}"
+          s << "    sudo cp -fv #{HOMEBREW_PREFIX}/opt/#{f.name}/*.plist #{destination}"
+          s << "    sudo launchctl load #{plist_link}"
+        else
+          s << "    launchctl unload #{plist_link}"
+          s << "    launchctl load #{plist_link}"
+        end
+      else
+        s << "To load #{f.name}:"
+        if f.plist_startup
+          s << "    sudo launchctl load #{plist_link}"
+        else
+          s << "    launchctl load #{plist_link}"
+        end
+      end
+
+      ohai 'Caveats', s
     end
   end
 
@@ -200,6 +259,8 @@ class FormulaInstaller
   end
 
   def build
+    FileUtils.rm Dir["#{HOMEBREW_LOGS}/#{f}/*"]
+
     @start_time = Time.now
 
     # 1. formulae can modify ENV, so we must ensure that each
@@ -236,7 +297,7 @@ class FormulaInstaller
       end
     end
 
-    ignore_interrupts do # the fork will receive the interrupt and marshall it back
+    ignore_interrupts(:quietly) do # the fork will receive the interrupt and marshall it back
       write.close
       Process.wait
       data = read.read
@@ -245,18 +306,14 @@ class FormulaInstaller
       raise "Suspicious installation failure" unless $?.success?
     end
 
-    # This is the installation receipt. The reason this comment is necessary
-    # is because some numpty decided to call the class Tab rather than
-    # the far more appropriate InstallationReceipt :P
-    Tab.for_install(f, args).write
+    raise "Empty installation" if Dir["#{f.prefix}/*"].empty?
+
+    Tab.for_install(f, args).write # INSTALL_RECEIPT.json
 
   rescue Exception => e
     ignore_interrupts do
       # any exceptions must leave us with nothing installed
-      if f.prefix.directory?
-        puts "One sec, just cleaning up..." if e.kind_of? Interrupt
-        f.prefix.rmtree
-      end
+      f.prefix.rmtree if f.prefix.directory?
       f.rack.rmdir_if_possible
     end
     raise
@@ -285,11 +342,11 @@ class FormulaInstaller
   end
 
   def install_plist
-    # Install a plist if one is defined
-    if f.startup_plist and not f.plist_path.exist?
-      f.plist_path.write f.startup_plist
-      f.plist_path.chmod 0644
-    end
+    return unless f.plist
+    # A plist may already exist if we are installing from a bottle
+    f.plist_path.unlink if f.plist_path.exist?
+    f.plist_path.write f.plist
+    f.plist_path.chmod 0644
   end
 
   def fix_install_names
@@ -323,7 +380,7 @@ class FormulaInstaller
 
   def pour
     fetched, downloader = f.fetch
-    f.verify_download_integrity fetched
+    f.verify_download_integrity fetched unless downloader.local_bottle_path
     HOMEBREW_CELLAR.cd do
       downloader.stage
     end
@@ -385,7 +442,7 @@ class FormulaInstaller
     return unless f.lib.directory?
 
     valid_extensions = %w(.a .dylib .framework .jnilib .la .o .so
-                          .jar .prl .pm)
+                          .jar .prl .pm .sh)
     non_libraries = f.lib.children.select do |g|
       next if g.directory?
       not valid_extensions.include? g.extname
@@ -431,25 +488,6 @@ class FormulaInstaller
   def audit_lib
     check_jars
     check_non_libraries
-  end
-
-  def check_m4
-    # Newer versions of Xcode don't come with autotools
-    return unless MacOS::Xcode.provides_autotools?
-
-    # If the user has added our path to dirlist, don't complain
-    return if File.open("/usr/share/aclocal/dirlist") do |dirlist|
-      dirlist.grep(%r{^#{HOMEBREW_PREFIX}/share/aclocal$}).length > 0
-    end rescue false
-
-    # Check for installed m4 files
-    if Dir[f.share+"aclocal/*.m4"].length > 0
-      opoo 'm4 macros were installed to "share/aclocal".'
-      puts "Homebrew does not append \"#{HOMEBREW_PREFIX}/share/aclocal\""
-      puts "to \"/usr/share/aclocal/dirlist\". If an autoconf script you use"
-      puts "requires these m4 macros, you'll need to add this path manually."
-      @show_summary_heading = true
-    end
   end
 end
